code/requirements.txt
azure-functions

streamlit==1.23.1
openai==0.27.7
matplotlib==3.7.1
plotly==5.14.1
scipy==1.10.1
scikit-learn==1.2.2
transformers==4.29.2
redis==4.5.5
python-dotenv==1.0.0
azure-ai-formrecognizer==3.2.1
azure-storage-blob==12.16.0
requests==2.31.0
tiktoken==0.4.0
azure-storage-queue==12.6.0
langchain==0.0.189
beautifulsoup4==4.12.2
streamlit-chat==0.0.2.2
fake-useragent==1.1.3
chardet==5.1.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.4.0a20230509004